Better handling of CPInfo bookkeeping
Each ColStruct record now has two CPInfo's: for one extent and for
overflow extent. These two records are set when extent is being added to
the list of extents to mark as invalid (actually, it is a list of
touched extents).

The conversion routine should handle conversion alongside updates to the
CPInfo structures.

This way we have unified machinery to keep the list of extents to mark
as invalid and to keep up-to-date CPRange information.